新增用户接口优化；私信功能完成

+++ ruoyi-modules/notify-service/src/main/java/com/ruoyi/notify/service/MessageService.java
@@ -10,7 +10,7 @@ import com.ruoyi.notify.domain.vo.ContactListVo;
 import java.util.List;
 
 public interface MessageService {
-    int send(SendDto dto);
+    long send(SendDto dto);
 
     List<ContactListVo> contactList(ListDto dto);
